Fixing Bug Kamar (Done)
